Simplify signature request workflow
